+++ src/vws/query.py
@@ -11,11 +11,16 @@ import requests
 from urllib3.filepost import encode_multipart_formdata
 from vws_auth_tools import authorization_header, rfc_1123_date
 
-from vws._result_codes import raise_for_result_code
-from vws.exceptions import (
-    ConnectionErrorPossiblyImageTooLarge,
+from vws.exceptions.cloud_reco_exceptions import (
+    AuthenticationFailure,
+    BadImage,
+    InactiveProject,
     MatchProcessing,
     MaxNumResultsOutOfRange,
+    RequestTimeTooSkewed,
+)
+from vws.exceptions.custom_exceptions import (
+    ConnectionErrorPossiblyImageTooLarge,
 )
 from vws.include_target_data import CloudRecoIncludeTargetData
 from vws.reports import QueryResult, TargetData
@@ -67,21 +72,22 @@ class CloudRecoService:
                 none (return no target_data), all (for all matched targets).
 
         Raises:
-            ~vws.exceptions.AuthenticationFailure: The client access key pair
-                is not correct.
-            ~vws.exceptions.MaxNumResultsOutOfRange: ``max_num_results`` is not
-                within the range (1, 50).
-            ~vws.exceptions.MatchProcessing: The given image matches a target
-                which was recently added, updated or deleted and Vuforia
-                returns an error in this case.
-            ~vws.exceptions.ProjectInactive: The project is inactive.
-            ~vws.exceptions.ConnectionErrorPossiblyImageTooLarge: The given
-                image is too large.
-            ~vws.exceptions.RequestTimeTooSkewed: There is an error with the
-                time sent to Vuforia.
-            ~vws.exceptions.BadImage: There is a problem with the given image.
-                For example, it must be a JPEG or PNG file in the grayscale or
-                RGB color space.
+            ~vws.exceptions.cloud_reco_exceptions.AuthenticationFailure: The
+                client access key pair is not correct.
+            ~vws.exceptions.cloud_reco_exceptions.MaxNumResultsOutOfRange:
+                ``max_num_results`` is not within the range (1, 50).
+            ~vws.exceptions.cloud_reco_exceptions.MatchProcessing: The given
+                image matches a target which was recently added, updated or
+                deleted and Vuforia returns an error in this case.
+            ~vws.exceptions.cloud_reco_exceptions.InactiveProject: The project
+                is inactive.
+            ~vws.exceptions.custom_exceptions.ConnectionErrorPossiblyImageTooLarge:
+                The given image is too large.
+            ~vws.exceptions.cloud_reco_exceptions.RequestTimeTooSkewed: There
+                is an error with the time sent to Vuforia.
+            ~vws.exceptions.cloud_reco_exceptions.BadImage: There is a problem
+                with the given image.  For example, it must be a JPEG or PNG
+                file in the grayscale or RGB color space.
 
         Returns:
             An ordered list of target details of matching targets.
@@ -137,10 +143,15 @@ class CloudRecoService:
         if 'No content to map due to end-of-input' in response.text:
             raise MatchProcessing(response=response)
 
-        raise_for_result_code(
-            response=response,
-            expected_result_code='Success',
-        )
+        result_code = response.json()['result_code']
+        if result_code != 'Success':
+            exception = {
+                'AuthenticationFailure': AuthenticationFailure,
+                'BadImage': BadImage,
+                'InactiveProject': InactiveProject,
+                'RequestTimeTooSkewed': RequestTimeTooSkewed,
+            }[result_code]
+            raise exception(response=response)
 
         result = []
         result_list = list(response.json()['results'])