+++ tests/test_vws.py
@@ -15,7 +15,7 @@ from mock_vws import MockVWS
 from mock_vws.database import VuforiaDatabase
 
 from vws import VWS, CloudRecoService
-from vws.exceptions import TargetProcessingTimeout
+from vws.exceptions.custom_exceptions import TargetProcessingTimeout
 from vws.reports import (
     DatabaseSummaryReport,
     TargetRecord,